Avoid signed index representation in gamepad axes configuration
Using a signed index for representing an axis index and a direction in
a single number may result in problematic behavior when the signed index
-0 (negative zero) is used. While JS IEEE 754 numbers include a
valid representation for negative zero, the sign is lost during
string conversion via JSON.stringify(), i.e. when the config is read
via the API server and not directly via the HTTP file server.

The updated configuration format avoids signed indices, but instead
allows for negating the horizontal or vertical axis if necessary.

See config/gamepads.xml for the documented of the updated gamepad
configuration schema.